Сделал разделение контента между пользователями системы

+++ ASP/publishing/publishing/Controllers/BookingsController.cs
@@ -2,24 +2,31 @@
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using publishing.Areas.Identity.Data;
 using publishing.Models;
 using publishing.Models.ViewModels;
 
 namespace publishing.Controllers
 {
+    [Authorize]
     public class BookingsController : Controller
     {
         private readonly PublishingDBContext _context;
+        private readonly UserManager<publishingUser> _userManager;
 
-        public BookingsController(PublishingDBContext context)
+        public BookingsController(PublishingDBContext context, UserManager<publishingUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         // GET: Bookings
+        [Authorize(Roles ="admin,manager")]
         public async Task<IActionResult> Index()
         {
             //BookingIndexViewModel model = new BookingIndexViewModel();
@@ -32,6 +39,7 @@ namespace publishing.Controllers
         }
 
         // GET: Bookings/Detail
+        [Authorize]
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null || _context.Bookings == null)
@@ -47,6 +55,9 @@ namespace publishing.Controllers
             {
                 return NotFound();
             }
+
+            if (!IsUserBooking(booking.Id))
+                return new StatusCodeResult(403);
 
             var products = (from bp in _context.BookingProducts.Include(bp => bp.Product).Include(bp => bp.Booking) where bp.BookingId == id select bp.Product).ToList();
             if (products == null)
@@ -67,6 +78,7 @@ namespace publishing.Controllers
         }
 
         // GET: Bookings/Create
+        [Authorize(Roles ="customer")]
         public IActionResult Create()
         {
             ViewData["PrintingHouseId"] = new SelectList(_context.PrintingHouses, "Id", "Name");
@@ -91,6 +103,7 @@ namespace publishing.Controllers
         }
 
         // GET: Bookings/Edit/5
+        [Authorize(Roles ="customer,admin")]
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null || _context.Bookings == null)
@@ -103,6 +116,10 @@ namespace publishing.Controllers
             {
                 return NotFound();
             }
+
+            if (!IsUserBooking(booking.Id))
+                return new StatusCodeResult(403);
+
             ViewData["PrintingHouseId"] = new SelectList(_context.PrintingHouses, "Id", "Name", booking.PrintingHouseId);
             return View(booking);
         }
@@ -144,6 +161,7 @@ namespace publishing.Controllers
         }
 
         // GET: Bookings/Delete/5
+        [Authorize(Roles ="customer,admin")]
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null || _context.Bookings == null)
@@ -158,6 +176,15 @@ namespace publishing.Controllers
             {
                 return NotFound();
             }
+
+            //var user = await _userManager.GetUserAsync(HttpContext.User);
+            //if (!await _userManager.IsInRoleAsync(user, "admin"))
+            //{
+            //    if (!IsUserBooking(booking.Id, user.Email))
+            //        return new StatusCodeResult(403);
+            //}
+            if (!IsUserBooking(booking.Id))
+                return new StatusCodeResult(403);
 
             return View(booking);
         }
@@ -186,6 +213,7 @@ namespace publishing.Controllers
           return (_context.Bookings?.Any(e => e.Id == id)).GetValueOrDefault();
         }
 
+        [Authorize(Roles ="customer")]
         public async Task<IActionResult> UnpinProduct(int? bookingId, int? productId)
         {
             if (bookingId == null || productId == null)
@@ -197,6 +225,9 @@ namespace publishing.Controllers
             if (booking == null || product == null)
                 return NotFound();
 
+            if (!IsUserBooking(booking.Id))
+                return new StatusCodeResult(403);
+
             if (_context.BookingProducts.Count(bp => bp.BookingId == bookingId) > 1)
             {
                 _context.BookingProducts.Remove(_context.BookingProducts.Single(bp => bp.ProductId == productId && bp.BookingId == bookingId));
@@ -206,9 +237,11 @@ namespace publishing.Controllers
                 costController.SetCostBooking(booking);
             }
 
-            return Redirect(Request.Headers["Referer"].ToString());
+             //return RedirectToAction("Details", new { id = bookingId });
+             return Redirect(Request.Headers["Referer"].ToString());
         }
 
+        [Authorize(Roles ="admin,manager")]
         public async Task<IActionResult> UnpinEmployee(int? bookingId, int? employeeId)
         {
             if (bookingId == null || employeeId == null)
@@ -224,11 +257,11 @@ namespace publishing.Controllers
                 _context.SaveChanges();
             }
 
-            return Redirect(Request.Headers["Referer"].ToString());
+            return Redirect(Request.Headers["Referer"].ToString()); // мб RedirectToAction
             //return View("Details", _context.Employees.Include(e => e.Bookings).Where(e => e.Id == employeeId).First());
         }
 
-
+        [Authorize(Roles ="admin,manager")]
         public IActionResult LinkEmployeeWithBooking(int? id)
         {
             if (id == null)
@@ -237,6 +270,9 @@ namespace publishing.Controllers
             Booking booking = _context.Bookings.Include(b => b.Employees).Single(b => b.Id == id);
             if (booking == null)
                 return NotFound();
+
+            if(booking.Status != "Выполняется")
+                return new StatusCodeResult(403);
 
             ViewBag.bookingId = id;
             return View(_context.Employees.Where(e=> !booking.Employees.Contains(e)));
@@ -265,6 +301,23 @@ namespace publishing.Controllers
             await _context.SaveChangesAsync();
 
             return RedirectToAction("Details", new { id = bookingId });
+        }
+
+        private bool IsUserBooking(int bookingId) 
+        {
+            //var user =  _userManager.GetUserAsync(HttpContext.User);
+
+            var user = _userManager.GetUserAsync(HttpContext.User);
+            //var user = await _userManager.GetUserAsync(HttpContext.User);
+            if ( _userManager.IsInRoleAsync(user.Result, "customer").Result)
+            {
+                var product = _context.BookingProducts.Where(bp => bp.BookingId == bookingId).Select(bp => bp.Product).First();
+                var customerProduct = _context.Products.Include(p => p.Customer).Single(p => p.Id == product.Id);
+
+                if (customerProduct.Customer.Email != user.Result.Email)
+                    return false;
+            }
+            return true;
         }
     }
 }
